- Remove packages - Refactor using statments - Move IparkingRepository interfave to Application project
