+++ ParkBuddy.Application/Handlers/QueryHandlers/GetParkingListHandler.cs
@@ -1,9 +1,8 @@
 using MediatR;
+using ParkBuddy.Application.Interfaces;
 using ParkBuddy.Application.Queries;
 using ParkBuddy.Contracts;
 using ParkBuddy.Contracts.Dtos;
-using ParkBuddy.Domain.Entities;
-using ParkBuddy.Domain.Repositories;
 
 namespace ParkBuddy.Application.Handlers.QueryHandlers
 {
