+++ ParkBuddy.Application/ParkBuddy.Application.csproj
@@ -2,13 +2,11 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ParkBuddy.Domain\ParkBuddy.Domain.csproj" />
-    <ProjectReference Include="..\ParkBuddy.Infrastructure\ParkBuddy.Infrastructure.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="FluentValidation.DependencyInjectionExtensions" Version="12.0.0-preview1" />
     <PackageReference Include="MediatR" Version="12.4.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.2" />
   </ItemGroup>
 
   <PropertyGroup>
