+++ ParkBuddy.Application/Validation/RegisterParkingDtoValidator.cs
@@ -1,4 +1,3 @@
-using System.Data;
 using FluentValidation;
 using ParkBuddy.Contracts.Dtos;
