+++ ParkBuddy.Server.sln
@@ -16,6 +16,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParkBuddy.Domain", "ParkBud
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParkBuddy.Infrastructure", "ParkBuddy.Infrastructure\ParkBuddy.Infrastructure.csproj", "{AFEE6D63-C792-493B-9F0B-1E6102F91C9A}"
+	ProjectSection(ProjectDependencies) = postProject
+		{B7C5A879-F51F-427C-A61C-39A5166B17F9} = {B7C5A879-F51F-427C-A61C-39A5166B17F9}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ParkBuddy.Services", "ParkBuddy.Services\ParkBuddy.Services.csproj", "{644D5674-D1C3-4560-8B1B-60F2119BD1B4}"
 EndProject
